wip: refactor: array index operator usage replaced by iterators
The goal is to eventually allow the underlying physical memory to be distributed rather than a huge contiguous block.

Iterators should open the door to using parts of the <algorithm> library, which may become useful as the container choice required grows in complexity.

This, however, undoubtfully adds a bunch of boilerplate code...